+++ Demo/ApplozicSwiftDemoTests/Tests/ALKConversationViewControllerSnapShotTests.swift
@@ -73,7 +73,9 @@ class ALKConversationViewControllerSnapshotTests: QuickSpec{
                 conversationVC.viewModel = ALKConversationViewModelMock(contactId: "demoUserId", channelKey: nil, localizedStringFileName: ALKConfiguration().localizedStringFileName)
                 conversationVC.beginAppearanceTransition(true, animated: false)
                 conversationVC.endAppearanceTransition()
-                conversationVC.updateDisplay(name: "demoDisplayName")
+                let contact = ALContact()
+                contact.displayName = "demoDisplayName"
+                conversationVC.updateDisplay(contact: contact, channel: nil)
                 navigationController = ALKBaseNavigationViewController(rootViewController: conversationVC)
             }