+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The changelog for [ApplozicSwift](https://github.com/AppLozic/ApplozicSwift). Al
 ---
 ### Enhancments
 - Updates Swift version to 4.2
+- Added Online/LastSeen status at navigation bar in conversation detail screen.
 
 2.2.0
 ---